pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["shivam <jhodshivam5122@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
django-filter = "^23.1"
django-split-settings = "^1.2.0"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pre-commit = "^3.2.2"
colorlog = "^6.7.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
